requirements.txt
flask==3.0.0
numpy==1.26.4
pillow==10.2.0
gunicorn==21.2.0
tensorflow-cpu==2.15.0